pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "python-framework"
dynamic = ["version"]
description = "Base Python application framework with PDM, Docker, and CI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Contributors" }]
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "openai>=1.59",
]

[project.scripts]
python-framework = "python_framework.cli:main"

[project.urls]
Repository = "https://github.com/example/python-framework"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.11",
    "pre-commit>=4.0",
    "bandit>=1.7",
    "pip-audit>=2.7",
    "cyclonedx-bom>=4.0",
]

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"
fallback_version = "0.1.0"

# PDM's environment.which() treats any command starting with "python" as the venv
# interpreter, so bare ``pdm run python-framework`` would become ``python hello``.
# This script delegates to ``python -m python_framework`` instead.
[tool.pdm.scripts]
python-framework = { cmd = "python -m python_framework {args}" }
prompt-example = { cmd = "python -m python_framework.examples.prompt_example {args}" }
structured-tutorial = { cmd = "python -m python_framework.examples.structured_outputs_tutorial.tutorial {args}" }

[tool.pdm.build]
package-dir = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-q",
    "--strict-markers",
    "--import-mode=importlib",
    "--cov=python_framework",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
]
filterwarnings = [
    "ignore:Module python_framework was previously imported, but not measured:coverage.exceptions.CoverageWarning",
]

[tool.coverage.run]
source_pkgs = ["python_framework"]
branch = true
omit = [
    "*/python_framework/__main__.py",
    "*/python_framework/examples/structured_outputs_tutorial/tutorial.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
packages = ["python_framework"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
warn_return_any = false

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
